requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
decouple==0.0.7
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
idna==3.10
jwt==1.4.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
